requirements.txt
anyio==3.5.0
asgiref==3.5.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
dnspython==2.2.1
fastapi==0.75.1
firebase-admin==5.2.0
google-api-core==2.7.2
google-api-python-client==2.44.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
msgpack==1.0.3
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pymongo==4.1.0
pyparsing==3.0.8
python-dotenv==0.20.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6

pip~=22.0.4
wheel~=0.37.1
Jinja2~=3.1.1
bson~=0.5.10
websockets~=10.2
ujson~=4.3.0
orjson~=3.6.7
uvloop~=0.16.0
PyYAML~=5.4.1
httptools~=0.2.0
watchgod~=0.8.2
itsdangerous~=2.1.2
python-multipart~=0.0.5
setuptools~=60.10.0